Allow to specify durations in weeks, months or years.
Make the --older-than option accept durations expressed as a number of
weeks (eg, '3w' for 21 days), months (eg, '2m' for 60 days), or years
(eg, '3y' for 1095 days).

For consistency, a 'd' suffix is also accepted to indicate a duration
in days, though it is already the default ('10' is the same as '10d').