Refactor the metrics generation code.
Move all the code related to the production of repository metrics to the
metrics module.

Store the metrics in a dedicated _MetricsReportSet object. Objects of
this class are generated by the MetricsReporter class, which is the
intended interface for the metrics module (along with the
MetricsFormatter class to write out the results).