metrics: Fix Markdown output.
The MarkdownMetricsFormatter was assuming that the first Report in a
ReportSet was always a report for the entire repository (selector
`all`), and was printing, for all the other Report objects, percentage
values computed relatively to that first Report.

Now that custom selectors can be used, there is no guarantee that the
first Report will always be a “all”-type Report, so we only print
percentage values if it happens to be the case.