Support multi-word values in filter.
GitHub labels may contain spaces, so we should accept spaces within the
expected values of `label` filters.

This introduces the need for a different method of specifying a custom
filter name, as `label:value AS name` is now interpreted as if `value AS
name` is the entire value of the `label` filter. Filter name can now be
specified with an equal sign instead, as in `label:value = name`.