Use slightly better-looking default selector names.
When a selector has no explicit name (using the `= name` syntax), derive
a default name from that selector’s syntax:

- `user:username` becomes `@username`;
- `team:teamname` becomes `teamname`;
- `label:labelname` becomes `labelname`.

Likewise in combined selectors. The derived names can be ambiguous
(there’s nothing to distinguish a team name from a label), but as they
are intended to be only cosmetic this is not an issue.