Fix handling of --from and --to option.
When a date (written as either YYYY-MM or YYYY-MM-DD) is passed to the
--from or the --to option, the YYYY bit is incorrectly interpreted as a
number of days, and the rest of the date is simply ignored.

This is because the regular expression in util.parse_duration does not
try to match the entire string, so YYYY-MM or YYYY-MM-DD actually
matches the expression.

The fix is to force a match against the entire string.