Obtain list of committers from the GitHub API.
We cannot easily derive the list of committers from the available data
on commits (because commits are associated to email addresses, not
GitHub users), so we use a dedicated function of the GitHub API to take
care of it for us. That's one more item to be fetched from GitHub and to
be stored in the local cache.