Expose commenters, contributors, and collaborators.
Add a new `commenters` property to the Repository object to provide the
list of all users who ever wrote something on the repository issues
tracker (i.e. users who opened a ticket or a PR, or who commented on a
ticket or a PR). The list is derived from already available information.

Change the meaning of the `contributors` property to get a list of all
users who ever did something on the repository (authored a commit.
opened a ticket/PR, commented on a ticket/PR). This is again derived
from already available information (`contributors` basically returns the
union of `committers` and `commenters`).

Add a new `collaborators` property with the former meaning attributed to
`contributors`: it returns the list of users registered as collaborators
on the repository, regardless of their actual contribution.

closes #16