Exclude 'github-actions[bot]' from user lists.
When listing committers and/or commenters, make sure to exclude
`github-actions[bot]` from the list.

Not only this is not informative for most (all?) purposes of GrainyHead,
but this causes issues when the list is later used to expand the
`user:*` selector (because the selector syntax does not allow the
`[bot]` part).