Use same "special names" for user: and team: selectors.
Use `user:*__contributors` (instead of `user:*CONTRIBUTORS`) to get
individual stats for contributors. This is for consistency with the
syntax expected by the team selector (which uses `team:__contributors`
to get grouped stats for all contributors).

Ditto for commenters, committers, and collaborators.

This uniformises the code to deal with those special groups, which may
now be all assimilated to special team names, as was already the case
for `__collaborators`.