Add type hints.
Add type hints whenever possible.

Mypy revealed that it is no longer possible to have a method that acts
as a starting property, so we simplify the CachePolicy class so that the
various REFRESH, NO_REFRESH, etc. members are simply directly assigned
class variables.

All the code involving calls to the GitHub API remains unchecked as we
do not have type hints for the corresponding Python module.